cloud_functions/etl_gcs_to_bq/requirements.txt
# Framework para Cloud Functions
functions-framework==3.*

# Librería de Google Cloud para BigQuery
google-cloud-bigquery>=3.0.0
